+++ braindecode/samplers/__init__.py
@@ -1,11 +1,18 @@
 """Classes to sample examples."""
 
-from .base import RecordingSampler, SequenceSampler, BalancedSequenceSampler
-from .ssl import RelativePositioningSampler
+from .base import (
+    RecordingSampler,
+    SequenceSampler,
+    BalancedSequenceSampler,
+    DistributedRecordingSampler,
+)
+from .ssl import RelativePositioningSampler, DistributedRelativePositioningSampler
 
 __all__ = [
     "RecordingSampler",
     "SequenceSampler",
     "BalancedSequenceSampler",
     "RelativePositioningSampler",
+    "DistributedRecordingSampler",
+    "DistributedRelativePositioningSampler",
 ]